Created a folder for the mock exams and moved the mock exam from 2022 to it
